requirements.txt
playwright>=1.42.0
python-dotenv>=1.0.0
openai>=1.0.0
